Put the libaries after the objects.
When linking the compiler (or at least gcc) looks at the files in the order they appear
on the commandline.  If the libaries are static and they are looked at first gcc doesn't
find anything useful and ingores everything in them.  Later on, gcc can't find the
refrences it needs in the object files.  Puting the libs after the objects causes gcc to
find everything it needs.